add view module
