+++ src/view/comment_view.rs
@@ -1,55 +1,10 @@
-use super::hn_client::*;
+use super::story_view::get_story_view;
+use crate::hn_client::*;
 use anyhow::Result;
-use cursive::{event::EventResult, traits::*, view::IntoBoxedView, views::*};
+use cursive::{traits::*, view::IntoBoxedView, views::*};
 use rayon::prelude::*;
 use regex::Regex;
 use std::time::{Duration, SystemTime};
-
-/// Construct a new Event view from a SelectView by adding
-/// event handlers for a key pressed
-fn construct_event_view<T: 'static>(view: SelectView<T>) -> OnEventView<SelectView<T>> {
-    // add "j" and "k" for moving down and up the story list
-    OnEventView::new(view)
-        .on_pre_event_inner('k', |s, _| {
-            let cb = s.select_up(1);
-            Some(EventResult::Consumed(Some(cb)))
-        })
-        .on_pre_event_inner('j', |s, _| {
-            let cb = s.select_down(1);
-            Some(EventResult::Consumed(Some(cb)))
-        })
-}
-
-/// Return a cursive's View from a story list
-pub fn get_story_view(stories: Vec<Story>, hn_client: &HNClient) -> impl IntoBoxedView {
-    let hn_client = hn_client.clone();
-    construct_event_view(
-        SelectView::new()
-            .with_all(stories.into_iter().enumerate().map(|(i, story)| {
-                (
-                    format!(
-                        "{}. {} (author: {}, {} comments, {} points)",
-                        i,
-                        story.title.clone().unwrap_or("unknown title".to_string()),
-                        story.author.clone().unwrap_or("unknown_user".to_string()),
-                        story.num_comments,
-                        story.points
-                    ),
-                    story,
-                )
-            }))
-            .on_submit(move |s, story| match get_comment_view(story, &hn_client) {
-                Err(err) => {
-                    log::error!("failed to construct comment view: {:#?}", err);
-                }
-                Ok(comment_view) => {
-                    s.pop_layer();
-                    s.add_layer(comment_view);
-                }
-            }),
-    )
-    .scrollable()
-}
 
 /// Parse a raw text from HN API to human-readable string
 fn format_hn_text(
@@ -121,7 +76,7 @@ fn parse_comment_text_list(comments: &Vec<Box<Comment>>, height: usize) -> Vec<(
 }
 
 /// Return a cursive's View from a comment list
-fn get_comment_view(story: &Story, hn_client: &HNClient) -> Result<impl IntoBoxedView> {
+pub fn get_comment_view(story: &Story, hn_client: &HNClient) -> Result<impl IntoBoxedView> {
     let hn_client = hn_client.clone();
 
     let comments = parse_comment_text_list(&story.get_comments(&hn_client)?, 0);
