+++ src/view/mod.rs
@@ -1,0 +1,2 @@
+pub mod comment_view;
+pub mod story_view;
